to_claude/requirements.txt
streamlit>=1.40.0
snowflake-connector-python>=4.3.0
anthropic>=0.40.0
pandas>=2.2.0
openpyxl>=3.1.0
cryptography>=46.0.0
openai>=1.0.0
audio-recorder-streamlit>=0.0.8
reportlab>=4.0.0
